Allow longer get key timeouts. Looking a commits to octvalidate, there seems to be a desire for this
git-svn-id: http://casvn/svn/octeon/bdk/branches/thunder@105226 c83883e8-9f28-0410-baf1-f2025811e16a

+++ libbdk-os/bdk-readline.c
@@ -846,7 +846,7 @@ static uint64_t gettime(void)
 }
 
 
-int bdk_readline_getkey(int timeout_us)
+int bdk_readline_getkey(uint64_t timeout_us)
 {
 #ifdef BDK_BUILD_HOST
     /* Switch to RAW terminal IO */
